워크스페이스 추가 컴포넌트 안으로 모달 관련 부분 이동
