+++ src/components/common/workspace/AddWorkspace.tsx
@@ -1,6 +1,10 @@
 import styled from '@emotion/styled';
 import plusLogo from '../../../resources/image/plusLogo.png';
 import { Workspace } from '@@types/index';
+import useAdminUser from '@hooks/useAdminUser';
+import { useRef, useState } from 'react';
+import InputModal from '../modal/InputModal';
+import { AddWorkspaceModalContent } from '../content/AddworkspaceModalContent';
 
 const AddWorkspaceContainer = styled.form`
   cursor: pointer;
@@ -15,18 +19,53 @@ const AddWorkspaceContainer = styled.form`
 
 type WrapperProps = {
   workspaces: Workspace[];
-  modalOpen: () => void;
 };
 
-const AddWorkspace = ({ workspaces, modalOpen }: WrapperProps) => {
+const AddWorkspace = ({ workspaces }: WrapperProps) => {
   const maxWorkspaceNum = 3;
+  const { createWorkspaces } = useAdminUser();
+  const workspaceNameRef = useRef<HTMLInputElement>(null);
+  const workspaceDescriptionRef = useRef<HTMLInputElement>(null);
+  const [modalOpen, setModalOpen] = useState<boolean>(false);
+
+  const createHandler = (e: React.FormEvent) => {
+    e.preventDefault();
+    e.stopPropagation();
+
+    const workspaceName = workspaceNameRef.current?.value;
+    if (!workspaceName) {
+      alert('workspace 이름을 입력해주세요');
+      return;
+    }
+
+    const workspaceDescription = workspaceDescriptionRef.current?.value;
+    if (!workspaceDescription) {
+      alert('workspace 설명 입력해주세요');
+      return;
+    }
+    console.log('hi');
+    createWorkspaces(workspaceName, workspaceDescription);
+    setModalOpen(false);
+  };
+
+  const closeModal = () => {
+    setModalOpen(false);
+  };
 
   if (workspaces.length >= maxWorkspaceNum) return null;
 
   return (
-    <AddWorkspaceContainer onClick={modalOpen}>
-      <img src={plusLogo} width={'51px'} height={'51px'}></img>
-    </AddWorkspaceContainer>
+    <>
+      <AddWorkspaceContainer onClick={() => setModalOpen(true)}>
+        <img src={plusLogo} width={'51px'} height={'51px'}></img>
+      </AddWorkspaceContainer>
+
+      {modalOpen && (
+        <InputModal closeModal={closeModal} createHandler={createHandler}>
+          <AddWorkspaceModalContent workspaceDescriptionRef={workspaceDescriptionRef} workspaceNameRef={workspaceNameRef} />
+        </InputModal>
+      )}
+    </>
   );
 };
